+++ .gitignore
@@ -11,3 +11,9 @@
 /build
 /captures
 .externalNativeBuild
+.idea/gradle.xml
+.idea/misc.xml
+.idea/runConfigurations.xml
+.idea/vcs.xml
+.idea/caches/gradle_models.ser
+.idea/codeStyles/Project.xml
